Convert JavaScriptCore to Modern Objective-C with ARC https://bugs.webkit.org/show_bug.cgi?id=251797
Reviewed by NOBODY (OOPS!).

This was a while in the making. This focuses on modernizing
and migrating the JavaScriptCore code to ARC, which will
require an API review, but nonetheless, with the code
upgraded and running under ARC,this is one incremental step
in the migration to ARC done.

* Source/JavaScriptCore/API/APICallbackFunction.h:
* Source/JavaScriptCore/API/JSAPIGlobalObject.mm:
* Source/JavaScriptCore/API/JSAPIWrapperObject.mm:
* Source/JavaScriptCore/API/JSContext.mm:
* Source/JavaScriptCore/API/JSContextInternal.h:
* Source/JavaScriptCore/API/JSManagedValue.h:
* Source/JavaScriptCore/API/JSManagedValue.mm:
* Source/JavaScriptCore/API/JSScript.h:
* Source/JavaScriptCore/API/JSScript.mm:
* Source/JavaScriptCore/API/JSScriptInternal.h:
* Source/JavaScriptCore/API/JSScriptSourceProvider.mm:
* Source/JavaScriptCore/API/JSValue.mm:
* Source/JavaScriptCore/API/JSVirtualMachine.mm:
* Source/JavaScriptCore/API/JSWrapperMap.mm:
* Source/JavaScriptCore/API/ObjCCallbackFunction.mm:
* Source/JavaScriptCore/API/tests/CurrentThisInsideBlockGetterTest.mm:
* Source/JavaScriptCore/API/tests/DateTests.mm:
* Source/JavaScriptCore/API/tests/Regress141275.mm:
* Source/JavaScriptCore/API/tests/Regress141809.mm:
* Source/JavaScriptCore/API/tests/testapi.mm:
* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/JavaScriptCore/inspector/ContentSearchUtilities.cpp:
* Source/JavaScriptCore/inspector/remote/cocoa/RemoteConnectionToTargetCocoa.mm:
* Source/JavaScriptCore/inspector/remote/cocoa/RemoteInspectorCocoa.mm:
* Source/JavaScriptCore/inspector/remote/cocoa/RemoteInspectorXPCConnection.mm:
* Source/JavaScriptCore/inspector/scripts/codegen/objc_generator.py:
* Source/JavaScriptCore/inspector/scripts/codegen/objc_generator_templates.py:
* Source/JavaScriptCore/inspector/scripts/tests/expected/type-declaration-object-type.json-result:
* Source/JavaScriptCore/inspector/scripts/tests/expected/type-requiring-runtime-casts.json-result:
* Source/JavaScriptCore/testmem/testmem.mm:

+++ Source/JavaScriptCore/API/glib/JSCException.cpp
@@ -391,9 +391,9 @@ char* jsc_exception_report(JSCException* exception)
     if (priv->sourceURI)
         report = g_string_append(report, priv->sourceURI.get());
     if (priv->lineNumber)
-        g_string_append_printf(report, ":%d", priv->lineNumber);
+        g_string_append_printf(report, ":%u", priv->lineNumber);
     if (priv->columnNumber)
-        g_string_append_printf(report, ":%d", priv->columnNumber);
+        g_string_append_printf(report, ":%u", priv->columnNumber);
     report = g_string_append_c(report, ' ');
     GUniquePtr<char> errorMessage(jsc_exception_to_string(exception));
     if (errorMessage)
@@ -402,7 +402,7 @@ char* jsc_exception_report(JSCException* exception)
 
     if (priv->backtrace) {
         GUniquePtr<char*> lines(g_strsplit(priv->backtrace.get(), "\n", 0));
-        for (unsigned i = 0; lines.get()[i]; ++i)
+        for (size_t i = 0; lines.get()[i]; ++i)
             g_string_append_printf(report, "  %s\n", lines.get()[i]);
     }
